// web-react/src/rpc/client.gen.ts
export interface Line {
  chordsOnly: boolean
  parts: Array<Part>
  annotation?: string
}

export interface Part {
  chord?: string
  text?: string
}